# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/contracts": {
      "post": {
        "operationId": "ContractsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "CREATE ",
        "tags": [
          "Contracts"
        ]
      },
      "get": {
        "operationId": "ContractsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 25,
              "default": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfContract",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Contract"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "GET ALL",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/{id}": {
      "get": {
        "operationId": "ContractsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "GET ONE",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/{id}/suspend": {
      "patch": {
        "operationId": "ContractsController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "SUSPEND CONTRACT",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/{id}/activate": {
      "patch": {
        "operationId": "ContractsController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateContractDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ACTIVATE",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/due/{id}/add_invoices": {
      "patch": {
        "operationId": "ContractsController_addInvoices",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddInvoicesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ADD INVOICES",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/due/{id}/update_invoice": {
      "patch": {
        "operationId": "ContractsController_updateInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "UPDATE INVOICE",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/due/{id}/remove_invoice": {
      "delete": {
        "operationId": "ContractsController_removeInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "REMOVE INVOICE",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/{id}/pdf": {
      "get": {
        "operationId": "ContractsController_getContractPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Contracts"
        ]
      }
    },
...
    "/billings/pay_dues": {
      "post": {
        "operationId": "BillingsController_payDue",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayDueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "PAY DUE",
        "tags": [
          "Billings"
        ]
      }
    },
    "/billings/request_payment": {
      "post": {
        "operationId": "BillingsController_requestPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayDueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "TEST REQUEST PAYMENT",
        "tags": [
          "Billings"
        ]
      }
    },
    "/billings/status_payment": {
      "post": {
        "operationId": "BillingsController_statusPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayCallbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "TEST STATUS PAYMENT",
        "tags": [
          "Billings"
        ]
      }
    },
...
  },
  "info": {
    "title": "LOCAPAY BACKEND",
    "description": "Documentation for the LOCAPAY API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://locapay-8f958cc17518.herokuapp.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegistryDates": {
        "type": "object",
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "updated_at",
          "deletedAt"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "MANAGER",
              "USER"
            ]
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "id": {
            "type": "number"
          },
          "registryDates": {
            "$ref": "#/components/schemas/RegistryDates"
          }
        },
        "required": [
          "name",
          "users",
          "id",
          "registryDates"
        ]
      },
      "Gallery": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "property": {
            "$ref": "#/components/schemas/Property"
          },
          "id": {
            "type": "number"
          },
          "registryDates": {
            "$ref": "#/components/schemas/RegistryDates"
          }
        },
        "required": [
          "url",
          "property",
          "id",
          "registryDates"
        ]
      },
      "Panorama": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "property": {
            "$ref": "#/components/schemas/Property"
          },
          "id": {
            "type": "number"
          },
          "registryDates": {
            "$ref": "#/components/schemas/RegistryDates"
          }
        },
        "required": [
          "url",
          "property",
          "id",
          "registryDates"
        ]
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "payment_type": {
            "type": "string",
            "enum": [
              "PAYMENT",
              "ADVANCE",
              "WITHDRAWAL"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "contract": {
            "$ref": "#/components/schemas/Contract"
          },
          "meta": {
            "type": "object"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "WALLET",
              "MTN"
            ]
          },
          "payment_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "FAILED",
              "SUCCESSFUL"
            ]
          },
          "id": {
            "type": "number"
          },
          "registryDates": {
            "$ref": "#/components/schemas/RegistryDates"
          }
        },
        "required": [
          "amount",
          "payment_type",
          "user",
          "contract",
          "meta",
          "payment_method",
          "payment_status",
          "id",
          "registryDates"
        ]
      },
...
      "Due": {
        "type": "object",
        "properties": {
          "amount_paid": {
            "type": "number"
          },
          "carry_over_amount": {
            "type": "number"
          },
          "due_date": {
            "format": "date-time",
            "type": "string"
          },
          "status_due": {
            "type": "string",
            "enum": [
              "WAITING",
              "IN_PROGRESS",
              "FINISHED"
            ]
          },
          "invoices": {
            "type": "object"
          },
          "contract": {
            "$ref": "#/components/schemas/Contract"
          },
          "annuities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annuity"
            }
          },
          "is_refunded": {
            "type": "boolean"
          },
          "id": {
            "type": "number"
          },
          "registryDates": {
            "$ref": "#/components/schemas/RegistryDates"
          }
        },
        "required": [
          "amount_paid",
          "carry_over_amount",
          "due_date",
          "status_due",
          "invoices",
          "contract",
          "annuities",
          "is_refunded",
          "id",
          "registryDates"
        ]
      },
      "Contract": {
        "type": "object",
        "properties": {
          "start_date": {
            "format": "date-time",
            "type": "string"
          },
          "end_date": {
            "format": "date-time",
            "type": "string"
          },
          "rent_price": {
            "type": "number"
          },
          "caution": {
            "type": "number"
          },
          "articles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                },
                "title": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "title",
                "content"
              ]
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "FINISHED",
              "SUSPENDED"
            ]
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          },
          "property": {
            "$ref": "#/components/schemas/Property"
          },
          "landlord": {
            "$ref": "#/components/schemas/User"
          },
          "tenant": {
            "$ref": "#/components/schemas/User"
          },
          "dues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Due"
            }
          },
          "is_gerant_locapay": {
            "type": "boolean"
          },
          "id": {
            "type": "number"
          },
          "registryDates": {
            "$ref": "#/components/schemas/RegistryDates"
          }
        },
        "required": [
          "start_date",
          "end_date",
          "rent_price",
          "caution",
          "articles",
          "status",
          "transactions",
          "property",
          "landlord",
          "tenant",
          "dues",
          "is_gerant_locapay",
          "id",
          "registryDates"
        ]
      },
      "Property": {
        "type": "object",
        "properties": {
          "to_sell": {
            "type": "boolean"
          },
          "video_url": {
            "type": "string"
          },
          "tarification": {
            "enum": [
              "DAILY",
              "MONTHLY"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "PARCEL",
              "HOUSE",
              "VILLA",
              "STUDIO",
              "APARTMENT",
              "APARTMENT_FURNITURE",
              "OFFICE",
              "BUILDING",
              "SHOP",
              "STORE"
            ],
            "type": "string"
          },
          "visit_price": {
            "type": "number"
          },
          "rent_price": {
            "type": "number"
          },
          "commission": {
            "type": "number"
          },
          "management_fee": {
            "type": "number"
          },
          "house_name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "district": {
            "type": "string"
          },
          "municipality": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "number_households": {
            "type": "number"
          },
          "number_living_rooms": {
            "type": "number"
          },
          "number_rooms": {
            "type": "number"
          },
          "number_bathrooms": {
            "type": "number"
          },
          "paint": {
            "enum": [
              "NO",
              "YES_CLIENT",
              "YES_OWNER"
            ],
            "type": "string"
          },
          "is_fence": {
            "type": "boolean"
          },
          "is_terace": {
            "type": "boolean"
          },
          "water_meter_type": {
            "enum": [
              "SONEB",
              "FORAGE"
            ],
            "type": "string"
          },
          "sanitary": {
            "enum": [
              "NO",
              "YES",
              "MIDDLE"
            ],
            "type": "string"
          },
          "electricity_meter_type": {
            "enum": [
              "PERSONAL",
              "DECOUNTER"
            ],
            "type": "string"
          },
          "electricity_personal_meter_type": {
            "enum": [
              "PREPAID",
              "POST_PREPAID"
            ],
            "type": "string"
          },
          "electricity_decounter_meter_rate": {
            "type": "number"
          },
          "month_advance": {
            "type": "number"
          },
          "caution": {
            "type": "number"
          },
          "water_drilling_rate": {
            "type": "number"
          },
          "is_prepaid": {
            "type": "boolean"
          },
          "is_active": {
            "type": "boolean"
          },
          "articles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                },
                "title": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "title",
                "content"
              ]
            }
          },
          "galleries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Gallery"
            }
          },
          "panorama": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Panorama"
            }
          },
          "visits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Visit"
            }
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "contracts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contract"
            }
          },
          "id": {
            "type": "number"
          },
          "registryDates": {
            "$ref": "#/components/schemas/RegistryDates"
          }
        },
        "required": [
          "to_sell",
          "video_url",
          "tarification",
          "type",
          "visit_price",
          "rent_price",
          "commission",
          "management_fee",
          "house_name",
          "description",
          "latitude",
          "longitude",
          "district",
          "municipality",
          "department",
          "number_households",
          "number_living_rooms",
          "number_rooms",
          "number_bathrooms",
          "paint",
          "is_fence",
          "is_terace",
          "water_meter_type",
          "sanitary",
          "electricity_meter_type",
          "electricity_personal_meter_type",
          "electricity_decounter_meter_rate",
          "month_advance",
          "caution",
          "water_drilling_rate",
          "is_prepaid",
          "is_active",
          "articles",
          "galleries",
          "panorama",
          "visits",
          "user",
          "owner",
          "contracts",
          "id",
          "registryDates"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "lastname": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "app_type": {
            "enum": [
              "LOCAPAY",
              "LOCAPAY_BUSINESS"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "NOT_VERIFIED",
              "PENDING",
              "ACCEPTED",
              "REJECTED"
            ],
            "type": "string"
          },
          "fcm_token": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "card_image": {
            "type": "string"
          },
          "card_number": {
            "type": "number"
          },
          "signature": {
            "type": "string"
          },
          "person_card": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "sexe": {
            "enum": [
              "M",
              "F"
            ],
            "type": "string"
          },
          "balance_mtn": {
            "type": "string"
          },
          "balance_moov": {
            "type": "string"
          },
          "wishlistedProperties": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Property"
            }
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          },
          "ownProperties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Property"
            }
          },
          "visits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Visit"
            }
          },
          "managedVisits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Visit"
            }
          },
          "ownerContracts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contract"
            }
          },
          "contracts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contract"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "id": {
            "type": "number"
          },
          "registryDates": {
            "$ref": "#/components/schemas/RegistryDates"
          }
        },
        "required": [
          "lastname",
          "firstname",
          "phone",
          "code",
          "password",
          "app_type",
          "status",
          "fcm_token",
          "image",
          "card_image",
          "card_number",
          "signature",
          "person_card",
          "reason",
          "sexe",
          "balance_mtn",
          "balance_moov",
          "wishlistedProperties",
          "roles",
          "properties",
          "transactions",
          "ownProperties",
          "visits",
          "managedVisits",
          "ownerContracts",
          "contracts",
          "notifications",
          "id",
          "registryDates"
        ]
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "itemsPerPage": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "currentPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          }
        },
        "required": [
          "itemsPerPage",
          "totalItems",
          "currentPage",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage"
        ]
      },
...
      "CreateContractDto": {
        "type": "object",
        "properties": {
          "tenant_id": {
            "type": "number"
          },
          "landlord_id": {
            "type": "number"
          },
          "property_id": {
            "type": "number"
          },
          "start_date": {
            "format": "date-time",
            "type": "string"
          },
          "is_gerant_locapay": {
            "type": "boolean"
          }
        },
        "required": [
          "tenant_id",
          "landlord_id",
          "property_id",
          "start_date",
          "is_gerant_locapay"
        ]
      },
      "ActivateContractDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "amount"
        ]
      },
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "title",
          "amount"
        ]
      },
      "AddInvoicesDto": {
        "type": "object",
        "properties": {
          "invoices": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateInvoiceDto"
            }
          }
        },
        "required": [
          "invoices"
        ]
      },
      "UpdateInvoiceDto": {
        "type": "object",
        "properties": {
          "invoice": {
            "$ref": "#/components/schemas/CreateInvoiceDto"
          },
          "invoice_id": {
            "type": "number"
          }
        },
        "required": [
          "invoice",
          "invoice_id"
        ]
      },
      "RemoveInvoiceDto": {
        "type": "object",
        "properties": {
          "invoice_id": {
            "type": "number"
          }
        },
        "required": [
          "invoice_id"
        ]
      },
      "PayDueDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "contract_id": {
            "type": "number"
          },
          "msisdn": {
            "type": "number"
          },
          "payment_type": {
            "enum": [
              "PAYMENT",
              "ADVANCE",
              "WITHDRAWAL"
            ],
            "type": "string"
          },
          "payment_method": {
            "enum": [
              "WALLET",
              "MTN"
            ],
            "type": "string"
          }
        },
        "required": [
          "amount",
          "contract_id",
          "msisdn",
          "payment_type",
          "payment_method"
        ]
      },
      "PayCallbackDto": {
        "type": "object",
        "properties": {
          "request_id_debit": {
            "type": "string"
          },
          "api_token": {
            "type": "string"
          }
        },
        "required": [
          "request_id_debit",
          "api_token"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}